Same for asset catalogs
